Revert "Merge branch 'main' into FireBall"
This reverts commit ce55b46c38a81274198bef8fc8936ae5b64423ee, reversing
changes made to a0246af471af8f317d326ee7491a101a258937c8.

+++ Engine/Source/Component/Mesh/Private/StaticMesh.cpp
@@ -89,7 +89,7 @@ void UStaticMesh::SetMaterial(int32 MaterialIndex, UMaterial* Material)
 
 int32 UStaticMesh::GetNumMaterials() const
 {
-	return static_cast<int32>(Materials.size());
+	return Materials.size();
 }
 
 const TArray<FMeshSection>& UStaticMesh::GetSections() const
